feat: qr 삭제 요청 구현

+++ src/components/qr/QrDropdown.tsx
@@ -1,4 +1,4 @@
-import { useNavigate, json } from "react-router-dom";
+import { useNavigate, json, redirect } from "react-router-dom";
 import { getAuthToken } from "../../util/auth";
 import style from "./QrDropdown.module.css";
 
@@ -11,25 +11,33 @@ const Dropdown: React.FC<{
     navigate(`/qr/${props.qrId}/edit`);
   };
   const onDeleteHandler = async () => {
-    const response = await fetch(
-      `https://qrecode-back.shop/qr/delete?idx=${props.qrId}`,
-      {
-        method: "DELETE",
-        headers: {
-          Authorization: `Bearer " + ${token}`,
-          "Content-Type": "application/json",
-        },
+    if (confirm("삭제 하시겠습니까?")) {
+      const response = await fetch(
+        `https://qrecode-back.shop/qr/delete?idx=${props.qrId}`,
+        {
+          method: "DELETE",
+          headers: {
+            Authorization: `Bearer " + ${token}`,
+            "Content-Type": "application/json",
+          },
+        }
+      );
+      if (!response.ok) {
+        throw new Error("QR 삭제에 실패했습니다.");
+      } else {
+        return navigate("/qr");
       }
-    );
-    if (!response.ok) {
-      throw json({ message: "QR 삭제에 실패했습니다." }, { status: 500 });
     }
   };
 
   return (
     <ul className={style.dropdown}>
-      <li className={style.li} onClick={onEditHandler}>수정</li>
-      <li className={style.li} onClick={onDeleteHandler}>삭제</li>
+      <li className={style.li} onClick={onEditHandler}>
+        수정
+      </li>
+      <li className={style.li} onClick={onDeleteHandler}>
+        삭제
+      </li>
     </ul>
   );
 };
